add activity/list, fix some minor stuff

+++ uu_onlyfitness_maing01-server/app/abl/activity-abl.js
@@ -12,24 +12,30 @@ const WARNINGS = {
       code: `${Errors.Create.UC_CODE}unsupportedKeys`,
     },
   },
-    Get: {
-      UnsupportedKeys: {
-        code: `${Errors.Get.UC_CODE}unsupportedKeys`,
-      },
+  
+  Get: {
+    UnsupportedKeys: {
+      code: `${Errors.Get.UC_CODE}unsupportedKeys`,
     },
+  },
 
-    Update: {
-      UnsupportedKeys: {
-        code: `${Errors.Update.UC_CODE}unsupportedKeys`,
-      },
+  Update: {
+    UnsupportedKeys: {
+      code: `${Errors.Update.UC_CODE}unsupportedKeys`,
     },
+  },
 
-    Delete: {
-      UnsupportedKeys: {
-        code: `${Errors.Delete.UC_CODE}unsupportedKeys`,
-      },
-    }
+  Delete: {
+    UnsupportedKeys: {
+      code: `${Errors.Delete.UC_CODE}unsupportedKeys`,
+    },
+  },
 
+  List: {
+    UnsupportedKeys: {
+      code: `${Errors.List.UC_CODE}unsupportedKeys`,
+    },
+  }
 };
 
 class ActivityAbl {
@@ -39,9 +45,61 @@ class ActivityAbl {
     this.dao = DaoFactory.getDao("activity");
   }
 
+  async list(awid, session, dtoIn) {
+
+    const validationResult = this.validator.validate("activityListDtoInType", dtoIn);
+
+    let uuAppErrorMap = {};
+    uuAppErrorMap = ValidationHelper.processValidationResult(
+      dtoIn,
+      validationResult,
+      uuAppErrorMap,
+      WARNINGS.List.UnsupportedKeys.code,
+      Errors.List.InvalidDtoIn
+    );
+
+    let filter = {
+      ownerId : session.getIdentity().getUuIdentity()
+    };
+
+    if(dtoIn.categoryId){
+      filter.categoryId = dtoIn.categoryId;
+    }
+
+    if(dtoIn.from && dtoIn.to){
+       filter.$and = [
+        {activityDate : { $gte: new Date(dtoIn.from) }}, 
+        {activityDate : { $lt: new Date(dtoIn.to) }} 
+       ]; 
+    } else if(dtoIn.from){
+      filter.activityDate = { $gte: new Date(dtoIn.from) };
+
+    } else if(dtoIn.to){
+      filter.activityDate = { $lt: new Date(dtoIn.to) };
+    }
+
+    let dtoOut;
+    try {
+      dtoOut = await this.dao.listByFilter(awid, filter, dtoIn.pageInfo, { "activityDate" : -1, "categoryId" : 1 });
+
+    } catch (e) {
+      if (e instanceof ObjectStoreError) {
+        throw new Errors.List.ActivityDaoListFailed({ uuAppErrorMap }, e);
+      }
+      throw e;
+    }
+
+    dtoOut.uuAppErrorMap = uuAppErrorMap;
+
+    return dtoOut;
+   
+  }
+
   async update(awid, session, dtoIn) {
 
     const validationResult = this.validator.validate("activityUpdateDtoInType", dtoIn);
+
+    //TODO validace existence category v category DAO    
 
     let uuAppErrorMap = ValidationHelper.processValidationResult(
       dtoIn,
@@ -51,13 +109,23 @@ class ActivityAbl {
       Errors.Update.InvalidDtoIn
     );
 
+    let entity = await this.dao.get(awid, dtoIn.id);
+    if (!entity){
+      throw new Errors.Update.ActivityNotPresent({ uuAppErrorMap });  
+    }
+
+    if (entity.ownerId != session.getIdentity().getUuIdentity()){
+      throw new Errors.Update.NotOwner({ uuAppErrorMap });  
+    }
+
     let dtoOut;
     try {
       dtoIn.awid = awid;
+      dtoIn.activityDate = new Date(dtoIn.activityDate);
       dtoOut = await this.dao.update(dtoIn);
     } catch (e) {
       if (e instanceof ObjectStoreError) {
-        throw new Errors.Create.ActivityDaoCreateFailed({ uuAppErrorMap }, e);
+        throw new Errors.Update.ActivityDaoUpdateFailed({ uuAppErrorMap }, e);
       }
       throw e;
     }
@@ -68,7 +136,7 @@ class ActivityAbl {
     
   }
 
-  async delete(awid,session, dtoIn) {
+  async delete(awid, session, dtoIn) {
     const validationResult = this.validator.validate("activityIdDtoInType", dtoIn);
 
     let uuAppErrorMap = {};
@@ -80,8 +148,16 @@ class ActivityAbl {
       Errors.Delete.InvalidDtoIn
     );
 
-    // todo - nejdriv get
-    await this.dao.remove(awid, dtoIn.id); // nevaliduje se, jestli vubec entita existuje
+    let entity = await this.dao.get(awid, dtoIn.id);
+    if (!entity){
+      throw new Errors.Delete.ActivityNotPresent({ uuAppErrorMap });  
+    }
+
+    if (entity.ownerId != session.getIdentity().getUuIdentity()){
+      throw new Errors.Delete.NotOwner({ uuAppErrorMap });  
+    }
+
+    await this.dao.remove(awid, dtoIn.id);
 
     let dtoOut = {
         awid,
@@ -94,9 +170,9 @@ class ActivityAbl {
   }
 
   async create(awid, session, dtoIn) {
-    console.log(">>>---");
-    console.log(dtoIn)
     const validationResult = this.validator.validate("activityCreateDtoInType", dtoIn);
+
+    //TODO validace existence category v category DAO
 
     let uuAppErrorMap = ValidationHelper.processValidationResult(
       dtoIn,
@@ -108,7 +184,6 @@ class ActivityAbl {
 
     let dtoOut;
     try {
-      console.log(session.getIdentity())
       dtoIn.awid = awid;
       dtoIn.ownerId = session.getIdentity().getUuIdentity();
       dtoIn.activityDate = new Date(dtoIn.activityDate);
